refactor(payroll): PayrollCalculator 1676→1509 lignes — extraction des agrégats de travail dans PayrollWorkInputs (Closes #5591)
Slice 1 du découpage du god-object : le concern « requêtes d'agrégation »
(collectWorkInputs / aggregateWorkInputs / sumApprovedLeaveDays — heures
sup, jours présents, congés payés/non payés, repli batch #2687) est extrait
dans PayrollWorkInputs, comportement IDENTIQUE (copie stricte).

- PayrollCalculator : délégation via la propriété workInputs (dépendance
  optionnelle au constructeur —  des tests intact) ;
  API publique collectWorkInputs() préservée (tests l'utilisent).
- Couverture : 480 tests verts (448 golden multi-pays + work inputs,
  coverage, edge) — filet de régression complet.
- Allowlist isolation modules mise à jour (mêmes dépendances, déplacées).
- Prochaines slices : calcul pur (computeSlipValues, 300 lignes),
  régularisation, indemnités.
